consistently use UPPER_SNAKE_CASE for constants (except in types whose interfaces mimic std types)

+++ include/donut/File.hpp
@@ -19,7 +19,7 @@ public:
 			: std::runtime_error(message) {}
 	};
 
-	static constexpr std::size_t NPOS = -1;
+	static constexpr std::size_t NPOS = static_cast<std::size_t>(-1);
 
 	static void createDirectory(const char* filepath);
